[ci] move test_max_errored_blocks as a first test
